pip-requirements.txt
#   Copyright 2023 University of Southampton
#   Dr Philip Basford
#   μ-VIS X-Ray Imaging Centre
#
#   Licensed under the Apache License, Version 2.0 (the "License");
#   you may not use this file except in compliance with the License.
#   You may obtain a copy of the License at
#
#       http://www.apache.org/licenses/LICENSE-2.0
#
#   Unless required by applicable law or agreed to in writing, software
#   distributed under the License is distributed on an "AS IS" BASIS,
#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
#   See the License for the specific language governing permissions and
#   limitations under the License.
wheel

django==2.2.20
cffi 
django-private-storage  
mysqlclient  
numpy 
pycparser 
pylibdmtx 
PyQRCode 
pytz 
pyvips 
sqlparse 
pypng 
sh 
django_mysql 
WeasyPrint 
ffmpeg-python 
pydrive2
python-magic  

#Check if https://github.com/django-polymorphic/django-polymorphic/pull/399 has been merged if not then need to download https://codeload.github.com/joshuamaxwell/django-polymorphic/zip/patch-1 and have in code directory.  May need to look in forks etc for a more up to date version
